extract relations logic to RelationsManager
