+++ postgres/datadog_checks/postgres/config_models/instance.py
@@ -18,6 +18,7 @@ class Relation(BaseModel):
         allow_mutation = False
 
     relation_name: Optional[str]
+    relation_regex: Optional[str]
     relation_schema: Optional[str]
     schemas: Optional[Sequence[str]]
