+++ postgres/datadog_checks/postgres/postgres.py
@@ -9,24 +9,19 @@ from six import iteritems
 
 from datadog_checks.base import AgentCheck
 from datadog_checks.postgres.metrics_cache import PostgresMetricsCache
+from datadog_checks.postgres.relationsmanager import (
+    IDX_METRICS,
+    LOCK_METRICS,
+    REL_METRICS,
+    SIZE_METRICS,
+    STATIO_METRICS,
+    RelationsManager,
+)
 from datadog_checks.postgres.statement_samples import PostgresStatementSamples
 from datadog_checks.postgres.statements import PostgresStatementMetrics
 
 from .config import PostgresConfig
-from .util import (
-    ALL_SCHEMAS,
-    CONNECTION_METRICS,
-    FUNCTION_METRICS,
-    IDX_METRICS,
-    LOCK_METRICS,
-    REL_METRICS,
-    REPLICATION_METRICS,
-    SIZE_METRICS,
-    STATIO_METRICS,
-    build_relations_filter,
-    fmt,
-    get_schema_field,
-)
+from .util import CONNECTION_METRICS, FUNCTION_METRICS, REPLICATION_METRICS, fmt, get_schema_field
 from .version_utils import V9, VersionUtils
 
 MAX_CUSTOM_RESULTS = 100
@@ -55,6 +50,7 @@ class PostgreSql(AgentCheck):
         self.metrics_cache = PostgresMetricsCache(self._config)
         self.statement_metrics = PostgresStatementMetrics(self, self._config)
         self.statement_samples = PostgresStatementSamples(self, self._config)
+        self._relations_manager = RelationsManager(self._config.relations)
         self._clean_state()
 
     def cancel(self):
@@ -87,38 +83,7 @@ class PostgreSql(AgentCheck):
             self._is_aurora = self._version_utils.is_aurora(self.db)
         return self._is_aurora
 
-    def _build_relations_config(self, yamlconfig):
-        """Builds a dictionary from relations configuration while maintaining compatibility"""
-        config = {}
-
-        for element in yamlconfig:
-            if isinstance(element, str):
-                config[element] = {'relation_name': element, 'schemas': [ALL_SCHEMAS]}
-            elif isinstance(element, dict):
-                if not ('relation_name' in element or 'relation_regex' in element):
-                    self.log.warning(
-                        "Parameter 'relation_name' or 'relation_regex' is required for relation element %s", element
-                    )
-                    continue
-                if 'relation_name' in element and 'relation_regex' in element:
-                    self.log.warning(
-                        "Expecting only of parameters 'relation_name', 'relation_regex' for relation element %s",
-                        element,
-                    )
-                    continue
-                schemas = element.get('schemas', [])
-                if not isinstance(schemas, list):
-                    self.log.warning("Expected a list of schemas for %s", element)
-                    continue
-                name = element.get('relation_name') or element['relation_regex']
-                config[name] = element.copy()
-                if len(schemas) == 0:
-                    config[name]['schemas'] = [ALL_SCHEMAS]
-            else:
-                self.log.warning('Unhandled relations config type: %s', element)
-        return config
-
-    def _run_query_scope(self, cursor, scope, is_custom_metrics, relations_config, cols, descriptors, is_relations):
+    def _run_query_scope(self, cursor, scope, is_custom_metrics, cols, descriptors, is_relations):
         if scope is None:
             return None
         if scope == REPLICATION_METRICS or not self.version >= V9:
@@ -132,7 +97,7 @@ class PostgreSql(AgentCheck):
             # if this is a relation-specific query, we need to list all relations last
             if is_relations:
                 schema_field = get_schema_field(descriptors)
-                relations_filter = build_relations_filter(relations_config, schema_field)
+                relations_filter = self._relations_manager.build_relations_filter(schema_field)
                 self.log.debug("Running query: %s with relations matching: %s", str(query), relations_filter)
                 cursor.execute(query.format(relations=relations_filter))
             else:
@@ -180,7 +145,7 @@ class PostgreSql(AgentCheck):
 
         return results
 
-    def _query_scope(self, cursor, scope, instance_tags, is_custom_metrics, relations_config):
+    def _query_scope(self, cursor, scope, instance_tags, is_custom_metrics):
         if scope is None:
             return None
         # build query
@@ -190,11 +155,9 @@ class PostgreSql(AgentCheck):
         # A descriptor is the association of a Postgres column name (e.g. 'schemaname')
         # to a tag name (e.g. 'schema').
         descriptors = scope['descriptors']
-        is_relations = scope['relation'] and len(relations_config) > 0
+        is_relations = scope['relation'] and len(self._relations_manager.config) > 0
 
-        results = self._run_query_scope(
-            cursor, scope, is_custom_metrics, relations_config, cols, descriptors, is_relations
-        )
+        results = self._run_query_scope(cursor, scope, is_custom_metrics, cols, descriptors, is_relations)
         if not results:
             return None
 
@@ -263,10 +226,8 @@ class PostgreSql(AgentCheck):
             metric_scope.append(self.metrics_cache.get_count_metrics())
 
         # Do we need relation-specific metrics?
-        relations_config = {}
         if self._config.relations:
             metric_scope += [LOCK_METRICS, REL_METRICS, IDX_METRICS, SIZE_METRICS, STATIO_METRICS]
-            relations_config = self._build_relations_config(self._config.relations)
 
         replication_metrics = self.metrics_cache.get_replication_metrics(self.version, self.is_aurora)
         if replication_metrics:
@@ -279,19 +240,19 @@ class PostgreSql(AgentCheck):
             metric_scope.append(replication_stats_metrics)
 
         cursor = self.db.cursor()
-        results_len = self._query_scope(cursor, db_instance_metrics, instance_tags, False, relations_config)
+        results_len = self._query_scope(cursor, db_instance_metrics, instance_tags, False)
         if results_len is not None:
             self.gauge("postgresql.db.count", results_len, tags=[t for t in instance_tags if not t.startswith("db:")])
 
-        self._query_scope(cursor, bgw_instance_metrics, instance_tags, False, relations_config)
-        self._query_scope(cursor, archiver_instance_metrics, instance_tags, False, relations_config)
+        self._query_scope(cursor, bgw_instance_metrics, instance_tags, False)
+        self._query_scope(cursor, archiver_instance_metrics, instance_tags, False)
 
         if self._config.collect_activity_metrics:
             activity_metrics = self.metrics_cache.get_activity_metrics(self.version)
-            self._query_scope(cursor, activity_metrics, instance_tags, False, relations_config)
+            self._query_scope(cursor, activity_metrics, instance_tags, False)
 
         for scope in list(metric_scope) + self._config.custom_metrics:
-            self._query_scope(cursor, scope, instance_tags, scope in self._config.custom_metrics, relations_config)
+            self._query_scope(cursor, scope, instance_tags, scope in self._config.custom_metrics)
 
         cursor.close()
