+++ postgres/datadog_checks/postgres/util.py
@@ -5,8 +5,6 @@ import string
 
 from datadog_checks.base import AgentCheck
 
-ALL_SCHEMAS = object()
-
 
 class PartialFormatter(string.Formatter):
     """Follows PEP3101, used to format only specified args in a string.
@@ -32,24 +30,6 @@ def get_schema_field(descriptors):
     for column, name in descriptors:
         if name == 'schema':
             return column
-
-
-def build_relations_filter(relations_config, schema_field):
-    """Build a WHERE clause filtering relations based on relations_config."""
-    relations_filter = []
-    for r in relations_config.values():
-        relation_filter = []
-        if r.get('relation_name'):
-            relation_filter.append("( relname = '{}'".format(r['relation_name']))
-        elif r.get('relation_regex'):
-            relation_filter.append("( relname ~ '{}'".format(r['relation_regex']))
-        if ALL_SCHEMAS not in r['schemas']:
-            schema_filter = ' ,'.join("'{}'".format(s) for s in r['schemas'])
-            relation_filter.append('AND {} = ANY(array[{}]::text[])'.format(schema_field, schema_filter))
-        relation_filter.append(')')
-        relations_filter.append(' '.join(relation_filter))
-
-    return ' OR '.join(relations_filter)
 
 
 fmt = PartialFormatter()
@@ -103,91 +83,6 @@ COMMON_ARCHIVER_METRICS = {
     'failed_count': ('postgresql.archiver.failed_count', AgentCheck.monotonic_count),
 }
 
-LOCK_METRICS = {
-    'descriptors': [
-        ('mode', 'lock_mode'),
-        ('locktype', 'lock_type'),
-        ('nspname', 'schema'),
-        ('datname', 'db'),
-        ('relname', 'table'),
-    ],
-    'metrics': {'lock_count': ('postgresql.locks', AgentCheck.gauge)},
-    'query': """
-SELECT mode,
-       locktype,
-       pn.nspname,
-       pd.datname,
-       pc.relname,
-       count(*) AS {metrics_columns}
-  FROM pg_locks l
-  JOIN pg_database pd ON (l.database = pd.oid)
-  JOIN pg_class pc ON (l.relation = pc.oid)
-  LEFT JOIN pg_namespace pn ON (pn.oid = pc.relnamespace)
- WHERE {relations}
-   AND l.mode IS NOT NULL
-   AND pc.relname NOT LIKE 'pg_%%'
- GROUP BY pd.datname, pc.relname, pn.nspname, locktype, mode""",
-    'relation': True,
-}
-
-REL_METRICS = {
-    'descriptors': [('relname', 'table'), ('schemaname', 'schema')],
-    'metrics': {
-        'seq_scan': ('postgresql.seq_scans', AgentCheck.rate),
-        'seq_tup_read': ('postgresql.seq_rows_read', AgentCheck.rate),
-        'idx_scan': ('postgresql.index_rel_scans', AgentCheck.rate),
-        'idx_tup_fetch': ('postgresql.index_rel_rows_fetched', AgentCheck.rate),
-        'n_tup_ins': ('postgresql.rows_inserted', AgentCheck.rate),
-        'n_tup_upd': ('postgresql.rows_updated', AgentCheck.rate),
-        'n_tup_del': ('postgresql.rows_deleted', AgentCheck.rate),
-        'n_tup_hot_upd': ('postgresql.rows_hot_updated', AgentCheck.rate),
-        'n_live_tup': ('postgresql.live_rows', AgentCheck.gauge),
-        'n_dead_tup': ('postgresql.dead_rows', AgentCheck.gauge),
-    },
-    'query': """
-SELECT relname,schemaname,{metrics_columns}
-  FROM pg_stat_user_tables
- WHERE {relations}""",
-    'relation': True,
-}
-
-IDX_METRICS = {
-    'descriptors': [('relname', 'table'), ('schemaname', 'schema'), ('indexrelname', 'index')],
-    'metrics': {
-        'idx_scan': ('postgresql.index_scans', AgentCheck.rate),
-        'idx_tup_read': ('postgresql.index_rows_read', AgentCheck.rate),
-        'idx_tup_fetch': ('postgresql.index_rows_fetched', AgentCheck.rate),
-    },
-    'query': """
-SELECT relname,
-       schemaname,
-       indexrelname,
-       {metrics_columns}
-  FROM pg_stat_user_indexes
- WHERE {relations}""",
-    'relation': True,
-}
-
-SIZE_METRICS = {
-    'descriptors': [('nspname', 'schema'), ('relname', 'table')],
-    'metrics': {
-        'pg_table_size(C.oid) as table_size': ('postgresql.table_size', AgentCheck.gauge),
-        'pg_indexes_size(C.oid) as index_size': ('postgresql.index_size', AgentCheck.gauge),
-        'pg_total_relation_size(C.oid) as total_size': ('postgresql.total_size', AgentCheck.gauge),
-    },
-    'relation': True,
-    'query': """
-SELECT
-  N.nspname,
-  relname,
-  {metrics_columns}
-FROM pg_class C
-LEFT JOIN pg_namespace N ON (N.oid = C.relnamespace)
-WHERE nspname NOT IN ('pg_catalog', 'information_schema') AND
-  nspname !~ '^pg_toast' AND
-  relkind = 'r' AND
-  {relations}""",
-}
 
 COUNT_METRICS = {
     'descriptors': [('schemaname', 'schema')],
@@ -280,27 +175,6 @@ WITH max_con AS (SELECT setting::float FROM pg_settings WHERE name = 'max_connec
 SELECT {metrics_columns}
   FROM pg_stat_database, max_con
 """,
-}
-
-STATIO_METRICS = {
-    'descriptors': [('relname', 'table'), ('schemaname', 'schema')],
-    'metrics': {
-        'heap_blks_read': ('postgresql.heap_blocks_read', AgentCheck.rate),
-        'heap_blks_hit': ('postgresql.heap_blocks_hit', AgentCheck.rate),
-        'idx_blks_read': ('postgresql.index_blocks_read', AgentCheck.rate),
-        'idx_blks_hit': ('postgresql.index_blocks_hit', AgentCheck.rate),
-        'toast_blks_read': ('postgresql.toast_blocks_read', AgentCheck.rate),
-        'toast_blks_hit': ('postgresql.toast_blocks_hit', AgentCheck.rate),
-        'tidx_blks_read': ('postgresql.toast_index_blocks_read', AgentCheck.rate),
-        'tidx_blks_hit': ('postgresql.toast_index_blocks_hit', AgentCheck.rate),
-    },
-    'query': """
-SELECT relname,
-       schemaname,
-       {metrics_columns}
-  FROM pg_statio_user_tables
- WHERE {relations}""",
-    'relation': True,
 }
 
 FUNCTION_METRICS = {
