+++ postgres/tox.ini
@@ -12,6 +12,8 @@ envdir =
 description =
     py{27,38}: e2e ready
 dd_check_style = true
+dd_check_types = true
+dd_mypy_args = --py2 datadog_checks/ tests/ --exclude '.*/config_models/.*\.py$'
 usedevelop = true
 platform = linux|darwin|win32
 passenv =
